genai/requirements.txt
fastapi
uvicorn
langchain
python-dotenv
requests
google-genai 
pydantic
weaviate-client